# services/api/cmd/server/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_gianghp123_SonaVoice_api_internal_core_enums.SessionStatus:
    enum:
    - pending
    - active
    - inactive
    - failed
    type: string
    x-enum-varnames:
    - SessionStatusPending
    - SessionStatusActive
    - SessionStatusInactive
    - SessionStatusFailed
  github_com_gianghp123_SonaVoice_api_internal_core_errors.AppError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_errors.AppError'
      meta:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.Meta'
      success:
        type: boolean
    type: object
...
  ? github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-array_github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res_SessionListItemRes
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res.SessionListItemRes'
        type: array
      error:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_errors.AppError'
      meta:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.Meta'
      success:
        type: boolean
    type: object
...
  ? github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res_CreateSessionRes
  : properties:
      data:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res.CreateSessionRes'
      error:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_errors.AppError'
      meta:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.Meta'
      success:
        type: boolean
    type: object
...
  ? github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res_SessionRes
  : properties:
      data:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res.SessionRes'
      error:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_errors.AppError'
      meta:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.Meta'
      success:
        type: boolean
    type: object
  github_com_gianghp123_SonaVoice_api_internal_core_response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res.CreateSessionRes:
    properties:
      id:
        type: string
      max_duration:
        type: integer
      webrtc_connection:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res.WebRTCConnectionRes'
    type: object
  github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res.IceConfig:
    properties:
      iceServers:
        items:
          $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res.IceServer'
        type: array
    type: object
  github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res.IceServer:
    properties:
      credential:
        description: for TURN servers
        type: string
      urls:
        items:
          type: string
        type: array
      username:
        description: for TURN servers
        type: string
    type: object
...
  github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res.SessionListItemRes:
    properties:
      created_at:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_enums.SessionStatus'
    type: object
  github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res.SessionRes:
    properties:
      created_at:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_enums.SessionStatus'
      user_id:
        type: string
    type: object
  github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res.WebRTCConnectionRes:
    properties:
      iceConfig:
        allOf:
        - $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res.IceConfig'
        description: 'only if enableDefaultIceServers: true'
      maxDuration:
        type: integer
      sessionId:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Session API
  version: "1.0"
paths:
...
  /sessions:
    get:
      consumes:
      - application/json
      description: Get paginated list of sessions belonging to the authenticated user
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-array_github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res_SessionListItemRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
      security:
      - BearerAuth: []
      summary: List user sessions
      tags:
      - session
    post:
      consumes:
      - application/json
      description: Create a new session and start a WebRTC connection with the speech
        service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res_CreateSessionRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
      security:
      - BearerAuth: []
      summary: Create new session and start connection
      tags:
      - session
  /sessions/{sessionId}:
    get:
      consumes:
      - application/json
      description: Get a single session by ID (must belong to the authenticated user)
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res_SessionRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
      security:
      - BearerAuth: []
      summary: Get session details
      tags:
      - session
  /sessions/{sessionId}/api/offer:
    patch:
      consumes:
      - application/json
      description: Proxy a WebRTC offer request to the speech service by session ID
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: WebRTC offer request body
        in: body
        name: body
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
      security:
      - BearerAuth: []
      summary: Proxy WebRTC offer
      tags:
      - session
    post:
      consumes:
      - application/json
      description: Proxy a WebRTC offer request to the speech service by session ID
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: WebRTC offer request body
        in: body
        name: body
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
      security:
      - BearerAuth: []
      summary: Proxy WebRTC offer
      tags:
      - session
  /sessions/{sessionId}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel a pending or active session, releasing all reserved quota
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
      security:
      - BearerAuth: []
      summary: Cancel a session
      tags:
      - session
...
  /sessions/{sessionId}/start:
    post:
      consumes:
      - application/json
      description: Start a WebRTC connection for a pending session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res_CreateSessionRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
      security:
      - BearerAuth: []
      summary: Start connection for an existing session
      tags:
      - session
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"